Compile benchmarks as separate modules
